module8-task1 | Перенос CITIES из моковых данных в константы, удаление лишних пробросов CITIES в пропсы
