+++ src/index.tsx
@@ -2,7 +2,6 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './components/common/app/app';
 import { HelmetProvider } from 'react-helmet-async';
-import {CITIES} from './mocks/cities';
 import ScrollToTop from './components/common/scroll-to-top/scroll-to-top';
 import {Provider} from 'react-redux';
 import {store} from '@/store';
@@ -21,7 +20,7 @@ root.render(
         <ScrollToTop />
         <Provider store={store}>
           <ToastContainer />
-          <App cities={CITIES} />
+          <App />
         </Provider>
       </BrowserRouter>
     </HelmetProvider>
